Clear branch
