+++ pcts_scrapers/items.py
@@ -6,7 +6,7 @@
 import scrapy
 
 
-class GenericScraperPaginationItem(scrapy.Item):
+class ScraperItem(scrapy.Item):
     # define the fields for your item here like:
     # name = scrapy.Field()
     source = scrapy.Field()
